Rendre générique la fonction qui récupère les fichiers des avis/saisines d'expert
